+++ backend/src/baserow/contrib/database/application_types.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import Any, Dict, Optional
+from typing import Any, Dict, Optional, List
 from zipfile import ZipFile
 
 from django.core.files.storage import Storage
@@ -60,28 +60,18 @@ class DatabaseApplicationType(ApplicationType):
     def export_safe_transaction_context(self, application) -> Atomic:
         return read_repeatable_single_database_atomic_transaction(application.id)
 
-    def export_serialized(
+    def export_tables_serialized(
         self,
-        database: Database,
+        tables: List[Table],
         files_zip: Optional[ZipFile] = None,
         storage: Optional[Storage] = None,
-    ) -> Dict[str, Any]:
+    ) -> List[Dict[str, Any]]:
         """
-        Exports the database application type to a serialized format that can later be
-        be imported via the `import_serialized`.
-        Call this function in a transaction_atomic(isolation_level=REPEATABLE_READ)
-        to ensure to read rows as they were at the beginning of the transaction,
-        independently from subsequent committed changes.
+        Exports the tables provided  to a serialized format that can later be
+        be imported via the `import_tables_serialized`.
         """
 
-        tables = database.table_set.all().prefetch_related(
-            "field_set",
-            "view_set",
-            "view_set__viewfilter_set",
-            "view_set__viewsort_set",
-        )
-
-        serialized_tables = []
+        serialized_tables: List[Dict[str, Any]] = []
         for table in tables:
             fields = table.field_set.all()
             serialized_fields = []
@@ -100,7 +90,7 @@ class DatabaseApplicationType(ApplicationType):
 
             model = table.get_model(fields=fields, add_dependencies=False)
             serialized_rows = []
-            table_cache = {}
+            table_cache: Dict[str, Any] = {}
             for row in model.objects.all():
                 serialized_row = DatabaseExportSerializedStructure.row(
                     id=row.id,
@@ -126,6 +116,27 @@ class DatabaseApplicationType(ApplicationType):
                     rows=serialized_rows,
                 )
             )
+        return serialized_tables
+
+    def export_serialized(
+        self,
+        database: Database,
+        files_zip: Optional[ZipFile] = None,
+        storage: Optional[Storage] = None,
+    ) -> Dict[str, Any]:
+        """
+        Exports the database application type to a serialized format that can later be
+        be imported via the `import_serialized`.
+        """
+
+        tables = database.table_set.all().prefetch_related(
+            "field_set",
+            "view_set",
+            "view_set__viewfilter_set",
+            "view_set__viewsort_set",
+        )
+
+        serialized_tables = self.export_tables_serialized(tables, files_zip, storage)
 
         serialized = super().export_serialized(database, files_zip, storage)
         serialized.update(
@@ -133,30 +144,16 @@ class DatabaseApplicationType(ApplicationType):
         )
         return serialized
 
-    def import_serialized(
-        self,
-        group: Group,
-        serialized_values: Dict[str, Any],
-        id_mapping: Dict[str, Any],
-        files_zip: Optional[ZipFile] = None,
-        storage: Optional[Storage] = None,
-        progress_builder: Optional[ChildProgressBuilder] = None,
-    ) -> Application:
-        """
-        Imports a database application exported by the `export_serialized` method.
-        """
-
-        tables = serialized_values.pop("tables")
-
-        child_total = (
-            # For the super application
-            1
+    def _ops_count_for_import_tables_serialized(
+        self, serialized_tables: List[Dict[str, Any]]
+    ) -> int:
+        return (
             +
             # Creating each table
-            len(tables)
+            len(serialized_tables)
             +
             # Creating each model table
-            len(tables)
+            len(serialized_tables)
             + sum(
                 [
                     # Inserting every field
@@ -169,52 +166,65 @@ class DatabaseApplicationType(ApplicationType):
                     len(table["rows"]) +
                     # After each field
                     len(table["fields"])
-                    for table in tables
+                    for table in serialized_tables
                 ]
             )
         )
-        progress = ChildProgressBuilder.build(progress_builder, child_total=child_total)
 
-        database = super().import_serialized(
-            group,
-            serialized_values,
-            id_mapping,
-            files_zip,
-            storage,
-            progress.create_child_builder(represents_progress=1),
-        )
+    def import_tables_serialized(
+        self,
+        database: Database,
+        serialized_tables: List[Dict[str, Any]],
+        id_mapping: Dict[str, Any],
+        files_zip: Optional[ZipFile] = None,
+        storage: Optional[Storage] = None,
+        progress_builder: Optional[ChildProgressBuilder] = None,
+    ) -> List[Table]:
+        """
+        Imports tables exported by the `export_tables_serialized` method.
+        This method has been created in order to import single tables or partial
+        applications. Beware to have all the tables needed in the
+        id_mapping["database_tables"] to make this works for link-row fields.
+        Look at `import_serialized` to know how to call this function.
+        """
+
+        child_total = self._ops_count_for_import_tables_serialized(serialized_tables)
+        progress = ChildProgressBuilder.build(progress_builder, child_total=child_total)
 
         if "database_tables" not in id_mapping:
             id_mapping["database_tables"] = {}
 
+        tables: List[Table] = []
+
         # First, we want to create all the table instances because it could be that
         # field or view properties depend on the existence of a table.
-        for table in tables:
-            table_object = Table.objects.create(
+        for serialized_table in serialized_tables:
+            table_instance = Table.objects.create(
                 database=database,
-                name=table["name"],
-                order=table["order"],
+                name=serialized_table["name"],
+                order=serialized_table["order"],
             )
-            id_mapping["database_tables"][table["id"]] = table_object.id
-            table["_object"] = table_object
-            table["_field_objects"] = []
+            id_mapping["database_tables"][serialized_table["id"]] = table_instance.id
+            serialized_table["_object"] = table_instance
+            serialized_table["_field_objects"] = []
+            tables.append(table_instance)
             progress.increment(state=IMPORT_SERIALIZED_IMPORTING)
 
         # Because view properties might depend on fields, we first want to create all
         # the fields.
         fields_excluding_reversed_linked_fields = []
         none_field_count = 0
-        for table in tables:
-            for field in table["fields"]:
+        for serialized_table in serialized_tables:
+            for field in serialized_table["fields"]:
                 field_type = field_type_registry.get(field["type"])
-                field_object = field_type.import_serialized(
-                    table["_object"], field, id_mapping
+                field_instance = field_type.import_serialized(
+                    serialized_table["_object"], field, id_mapping
                 )
 
-                if field_object:
-                    table["_field_objects"].append(field_object)
+                if field_instance:
+                    serialized_table["_field_objects"].append(field_instance)
                     fields_excluding_reversed_linked_fields.append(
-                        (field_type, field_object)
+                        (field_type, field_instance)
                     )
                 else:
                     none_field_count += 1
@@ -227,30 +237,30 @@ class DatabaseApplicationType(ApplicationType):
 
         # Now that the all tables and fields exist, we can create the views and create
         # the table schema in the database.
-        for table in tables:
-            for view in table["views"]:
+        for serialized_table in serialized_tables:
+            for view in serialized_table["views"]:
                 view_type = view_type_registry.get(view["type"])
                 view_type.import_serialized(
-                    table["_object"], view, id_mapping, files_zip, storage
+                    serialized_table["_object"], view, id_mapping, files_zip, storage
                 )
                 progress.increment(state=IMPORT_SERIALIZED_IMPORTING)
 
             # We don't need to create all the fields individually because the schema
             # editor can handle the creation of the table schema in one go.
             with safe_django_schema_editor() as schema_editor:
-                model = table["_object"].get_model(
-                    fields=table["_field_objects"],
+                table_model = serialized_table["_object"].get_model(
+                    fields=serialized_table["_field_objects"],
                     field_ids=[],
                     managed=True,
                     add_dependencies=False,
                 )
-                table["_model"] = model
-                schema_editor.create_model(model)
+                serialized_table["_model"] = table_model
+                schema_editor.create_model(table_model)
 
                 # The auto_now_add and auto_now must be disabled for all fields
                 # because the export contains correct values and we don't want them
                 # to be overwritten when importing.
-                for model_field in table["_model"]._meta.get_fields():
+                for model_field in serialized_table["_model"]._meta.get_fields():
                     if hasattr(model_field, "auto_now_add"):
                         model_field.auto_now_add = False
 
@@ -261,12 +271,14 @@ class DatabaseApplicationType(ApplicationType):
 
         # Now that everything is in place we can start filling the table with the rows
         # in an efficient matter by using the bulk_create functionality.
-        for table in tables:
-            model = table["_model"]
-            field_ids = [field_object.id for field_object in table["_field_objects"]]
+        for serialized_table in serialized_tables:
+            table_model = serialized_table["_model"]
+            field_ids = [
+                field_object.id for field_object in serialized_table["_field_objects"]
+            ]
             rows_to_be_inserted = []
 
-            for row in table["rows"]:
+            for row in serialized_table["rows"]:
                 created_on = row.get("created_on")
                 updated_on = row.get("updated_on")
 
@@ -280,14 +292,14 @@ class DatabaseApplicationType(ApplicationType):
                 else:
                     updated_on = timezone.now()
 
-                row_object = model(
+                row_object = table_model(
                     id=row["id"],
                     order=row["order"],
                     created_on=created_on,
                     updated_on=updated_on,
                 )
 
-                for field in table["fields"]:
+                for field in serialized_table["fields"]:
                     field_type = field_type_registry.get(field["type"])
                     new_field_id = id_mapping["database_fields"][field["id"]]
                     field_name = f'field_{field["id"]}'
@@ -310,23 +322,23 @@ class DatabaseApplicationType(ApplicationType):
 
                 rows_to_be_inserted.append(row_object)
                 progress.increment(
-                    state=f"{IMPORT_SERIALIZED_IMPORTING_TABLE}{table['id']}"
+                    state=f"{IMPORT_SERIALIZED_IMPORTING_TABLE}{serialized_table['id']}"
                 )
 
             # We want to insert the rows in bulk because there could potentially be
             # hundreds of thousands of rows in there and this will result in better
             # performance.
             for chunk in grouper(512, rows_to_be_inserted):
-                model.objects.bulk_create(chunk, batch_size=512)
+                table_model.objects.bulk_create(chunk, batch_size=512)
                 progress.increment(
                     len(chunk),
-                    state=f"{IMPORT_SERIALIZED_IMPORTING_TABLE}{table['id']}",
+                    state=f"{IMPORT_SERIALIZED_IMPORTING_TABLE}{serialized_table['id']}",
                 )
 
             # When the rows are inserted we keep the provide the old ids and because of
             # that the auto increment is still set at `1`. This needs to be set to the
             # maximum value because otherwise creating a new row could later fail.
-            sequence_sql = connection.ops.sequence_reset_sql(no_style(), [model])
+            sequence_sql = connection.ops.sequence_reset_sql(no_style(), [table_model])
             with connection.cursor() as cursor:
                 cursor.execute(sequence_sql[0])
 
@@ -341,5 +353,48 @@ class DatabaseApplicationType(ApplicationType):
         # Add the remaining none fields that we must not include in the import
         # because they were for example reversed link row fields.
         progress.increment(none_field_count, state=IMPORT_SERIALIZED_IMPORTING)
+        return tables
+
+    def import_serialized(
+        self,
+        group: Group,
+        serialized_values: Dict[str, Any],
+        id_mapping: Dict[str, Any],
+        files_zip: Optional[ZipFile] = None,
+        storage: Optional[Storage] = None,
+        progress_builder: Optional[ChildProgressBuilder] = None,
+    ) -> Application:
+        """
+        Imports a database application exported by the `export_serialized` method.
+        """
+
+        serialized_tables = serialized_values.pop("tables")
+        database_progress, table_progress = 1, 99
+        progress = ChildProgressBuilder.build(
+            progress_builder, child_total=database_progress + table_progress
+        )
+
+        application = super().import_serialized(
+            group,
+            serialized_values,
+            id_mapping,
+            files_zip,
+            storage,
+            progress.create_child_builder(represents_progress=database_progress),
+        )
+
+        database = application.specific
+
+        if not serialized_tables:
+            progress.increment(state=IMPORT_SERIALIZED_IMPORTING, by=table_progress)
+        else:
+            self.import_tables_serialized(
+                database,
+                serialized_tables,
+                id_mapping,
+                files_zip,
+                storage,
+                progress.create_child_builder(represents_progress=table_progress),
+            )
 
         return database