+++ web-frontend/modules/database/components/sidebar/SidebarItem.vue
@@ -42,6 +42,17 @@
           </a>
         </li>
         <li>
+          <SidebarDuplicateTableContextItem
+            :database="database"
+            :table="table"
+            :disabled="deleteLoading"
+            @table-duplicated="
+              $refs.context.hide()
+              selectTable(database, $event.table)
+            "
+          ></SidebarDuplicateTableContextItem>
+        </li>
+        <li>
           <a
             :class="{ 'context__menu-item--loading': deleteLoading }"
             @click="deleteTable()"
@@ -65,10 +76,15 @@
 import { notifyIf } from '@baserow/modules/core/utils/error'
 import ExportTableModal from '@baserow/modules/database/components/export/ExportTableModal'
 import WebhookModal from '@baserow/modules/database/components/webhook/WebhookModal'
+import SidebarDuplicateTableContextItem from '@baserow/modules/database/components/sidebar/table/SidebarDuplicateTableContextItem'
 
 export default {
   name: 'SidebarItem',
-  components: { ExportTableModal, WebhookModal },
+  components: {
+    ExportTableModal,
+    WebhookModal,
+    SidebarDuplicateTableContextItem,
+  },
   props: {
     database: {
       type: Object,