+++ web-frontend/modules/database/store/table.js
@@ -100,6 +100,22 @@ export const actions = {
     return data
   },
   /**
+   * Fetches one table for the authenticated user.
+   */
+  async fetch({ commit, dispatch }, { database, tableId }) {
+    commit('SET_LOADING', true)
+
+    try {
+      const { data } = await TableService(this.$client).get(tableId)
+      dispatch('forceCreate', { database, data })
+      commit('SET_LOADING', false)
+      return data
+    } catch (error) {
+      commit('SET_LOADING', false)
+      throw error
+    }
+  },
+  /**
    * Forcefully create an item in the store without making a call to the server.
    */
   forceCreate({ commit }, { database, data }) {