+++ changelog.md
@@ -22,6 +22,7 @@ For example:
 * Added option to use view's filters and sorting when listing rows. [#190](https://gitlab.com/bramw/baserow/-/issues/190)
 * Fixed bug with 404 middleware returning different 404 error messages based on the endpoint.
 * Made it possible to import data into an existing table. [#342](https://gitlab.com/bramw/baserow/-/issues/342)
+* Tables can now be duplicated. [#961](https://gitlab.com/bramw/baserow/-/issues/961)
 
 ### Bug Fixes